Fixed build.
Caused by Failed test that I swear was working pre-commit, but two
builds failed from the same test, so *shrug*. Possibly related to Json,
or Json-String comparison being fiddly. Decided to reimplement
AssertEquivalentLists to be a bit more explicit. It's a bit more work
and maintenance, but the code is much more straight forward.